requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<3.16
python-dotenv>=1.0
openai>=1.43
faiss-cpu>=1.7.4
numpy>=1.26
scikit-learn>=1.4
python-pptx>=0.6.23
PyPDF2>=3.0
python-docx>=1.0
tiktoken>=0.7
pydantic>=2.8
requests>=2.32
drf-spectacular>=0.27
django-cors-headers>=4.4
dj-database-url>=2.1
psycopg[binary]>=3.2
pgvector>=0.2
agentscope>=0.1.0
uvicorn[standard]>=0.30
